**************************************************************************** ** Merged Source Code from Development trunk, for the Spring 2011 Release ** **************************************************************************** o Development Trunk (Revision: 4444):  https://81.169.133.124/source/ja2/branches/Wanne/JA2%201.13%20MP o This trunk (the official 1.13 source trunk) will only be used for fixing bugs. No new features will be added here.   New Features are only included in the Development trunk o !!! After we have fixed a bug here in this trunk, we should MANUALLY merge the bugfix in the development trunk !!!
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@4446 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Laptop/XML_ShippingDestinations.cpp
@@ -160,7 +160,7 @@ destinationEndElementHandle(void *userData, const XML_Char *name)
 		else if(strcmp(name, "sGridNo") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->tempDest.sGridNo = (UINT16) atoi(pData->szCharData);
+			pData->tempDest.sGridNo = (UINT32) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "uiIndex") == 0)
 		{
